machine_learning-kNN
This file uses a k-Nearest Neighbors (k-NN) classifier and NumPy, Pandas, and Sci-Kit Learn to make predictions on the outcomes of eSports. We gathered data from the Riot Games API, split the data into training and testing, built the model, and tested the model. We had an accuracy of around 98 percent.